Stops test drive exploit SHOUTOUT@BUFFSOVERNEXUS
There is an exploit where people will test drive a vehicle and have someone else hop in the vehicle with them. There are additions on client side to ensure they are removed from vehicle prior to ending the test so they can't do the exploit. This addition is by BuffsOverNexus not me. https://github.com/BuffsOverNexus